Add files via upload
I included a few examples that illustrate how the libraries work and also two .cpp that I've chosen out of many (every .cpp is a system under study). I included the system of 2 qubits coupled with each other under closed system dynamics and also the system of two qubits coupled through a bosonic mode and in contact with external baths.